Combine the unit and samples travis scripts into a single job.
The unit test is substantially faster than all of the other jobs. We
have so many jobs scheduled that they don’t run in parallel and it’s 
not uncommon for a couple to be started only after all of the others
have finished. There’s a substantial amount of overlap between the
unit and samples jobs. Finally most of the samples script right now is
just waiting for firebase to run tests, so we might as well make use
of that time to run the unit tests.